+++ src/CostFunctions.cpp
@@ -73,7 +73,7 @@ Eigen::Matrix<double, 1, 6> PhotometricError::computeAnalyticalJacobian(const Me
                                                                         const Eigen::Matrix3d& R,
                                                                         const Eigen::Vector3d& t,
                                                                         const cv::Mat& image,
-                                                                        int u, int v) const {
+                                                                        double u, double v) const {
     Eigen::Matrix<double, 1, 6> J;
     J.setZero();
